feat: work on tournaments
