+++ apps/admin/components/toolbar/ToolbarContainer.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="ml-4 flex items-center gap-2">
+    <slot />
+  </div>
+</template>
