+++ packages/core/package.json
@@ -19,7 +19,7 @@
     "@nuxt/image": "^1.8.1",
     "@nuxt/fonts": "^0.8.0",
     "@nuxt/test-utils": "^3.14.2",
-    "@nuxt/ui": "^2.18.4",
+    "@nuxt/ui": "^2.19.2",
     "tailwindcss": "^3.4.11",
     "typescript": "^5.6.2"
   }
